fix(agent): isolate model metadata from client hooks
