+++ lib/hooks/use-agent-model.ts
@@ -1,3 +1,5 @@
+'use client'
+
 /**
  * useAgentModel Hook
  *
@@ -6,7 +8,15 @@
  */
 
 import { useEffect, useMemo, useState } from 'react'
+import {
+  AGENT_MODELS,
+  isKnownModel,
+  type ModelPricing,
+  type OpenAIModel,
+} from '@/lib/ai/agent-models'
 import { formatCompactNumber } from '@/lib/format-number'
+
+export type { OpenAIModel } from '@/lib/ai/agent-models'
 
 /**
  * LocalStorage key for model selection
@@ -21,87 +31,6 @@ const DEFAULT_MODEL = 'openrouter/free'
  */
 export function formatTokenCount(count: number): string {
   return formatCompactNumber(count)
-}
-
-/**
- * Available agent models.
- *
- * Keep `name` identical to the model code so the UI always shows
- * the exact provider/model identifier chosen by the user.
- */
-export const AGENT_MODELS = {
-  'openrouter/free': {
-    name: 'openrouter/free',
-    description:
-      'OpenRouter auto-router: picks a working free tool-capable model (default)',
-    contextLength: 200000,
-  },
-  'openrouter/auto': {
-    name: 'openrouter/auto',
-    description: 'OpenRouter auto-router: picks the best model (paid)',
-    contextLength: 2000000,
-  },
-  'z-ai/glm-4.5-air:free': {
-    name: 'z-ai/glm-4.5-air:free',
-    description: 'Z.AI GLM 4.5 Air, free tier',
-    contextLength: 131072,
-  },
-  'arcee-ai/trinity-large-preview:free': {
-    name: 'arcee-ai/trinity-large-preview:free',
-    description: 'Arcee Trinity Large Preview, free tier',
-    contextLength: 131000,
-  },
-  'qwen/qwen3-coder:free': {
-    name: 'qwen/qwen3-coder:free',
-    description: 'Qwen3 Coder, free tier, 1M context',
-    contextLength: 1048576,
-  },
-  'qwen/qwen3-next-80b-a3b-instruct:free': {
-    name: 'qwen/qwen3-next-80b-a3b-instruct:free',
-    description: 'Qwen3 Next 80B Instruct, free tier',
-    contextLength: 262144,
-  },
-  'openai/gpt-oss-120b:free': {
-    name: 'openai/gpt-oss-120b:free',
-    description: 'OpenAI GPT-OSS 120B, free tier',
-    contextLength: 131072,
-  },
-  'openai/gpt-oss-20b:free': {
-    name: 'openai/gpt-oss-20b:free',
-    description: 'OpenAI GPT-OSS 20B, free tier',
-    contextLength: 131072,
-  },
-  'meta-llama/llama-3.3-70b-instruct:free': {
-    name: 'meta-llama/llama-3.3-70b-instruct:free',
-    description: 'Meta Llama 3.3 70B Instruct, free tier',
-    contextLength: 131072,
-  },
-  'google/gemma-4-31b-it:free': {
-    name: 'google/gemma-4-31b-it:free',
-    description: 'Google Gemma 4 31B Instruct, free tier',
-    contextLength: 262144,
-  },
-  'minimax/minimax-m2.7': {
-    name: 'minimax/minimax-m2.7',
-    description: 'MiniMax production model (paid)',
-    contextLength: 200000,
-    pricing: {
-      inputPerMillion: 0.5,
-      outputPerMillion: 1.5,
-    },
-  },
-} as const
-
-/** OpenAI-compatible model identifier — any string is valid (custom models supported) */
-export type OpenAIModel = string
-
-/**
- * Check whether a model ID is in the known AGENT_MODELS list
- */
-export function isKnownModel(
-  model: string
-): model is keyof typeof AGENT_MODELS {
-  return model in AGENT_MODELS
 }
 
 /**
@@ -147,14 +76,6 @@ function saveModel(model: OpenAIModel): void {
   } catch {
     // localStorage may be disabled
   }
-}
-
-/**
- * Pricing information for a model (USD per 1M tokens)
- */
-export interface ModelPricing {
-  inputPerMillion: number
-  outputPerMillion: number
 }
 
 /**
